trivial: Fix a build warning from gcc 15.1.1
(cherry picked from commit 2fd2152a8092677f49273118fa300ec2833a3ac2)

+++ libfwupd/fwupd-client.c
@@ -840,7 +840,7 @@ fwupd_client_curl_new(FwupdClient *self, GError **error)
 	(void)curl_easy_setopt(helper->curl, CURLOPT_CAINFO, "ca-bundle.crt");
 #endif
 #if CURL_AT_LEAST_VERSION(7, 71, 0)
-	(void)curl_easy_setopt(helper->curl, CURLOPT_SSL_OPTIONS, CURLSSLOPT_NATIVE_CA);
+	(void)curl_easy_setopt(helper->curl, CURLOPT_SSL_OPTIONS, (glong)CURLSSLOPT_NATIVE_CA);
 #endif
 
 	/* this disables the double-compression of the firmware.xml.gz file */
